+++ source/Vector.cpp
@@ -1,7 +1,7 @@
 #include "..\\header\Vector.h"
 
 template <class type>
-Vector<type>::Vector(ll sz) : size(sz)
+Vector<type>::Vector(int sz) : size(sz)
 {
     if (size < 0)
         size = 1;
@@ -21,20 +21,20 @@ void Vector<type>::expand_capacity()
 {
     capacity *= 2;
     type *arr2 = new type[capacity]{};
-    for (ll i(0); i < size; ++i)
+    for (int i(0); i < size; ++i)
         arr2[i] = arr[i];
     swap(arr, arr2);
     delete[] arr2;
 }
 
 template <class type>
-ll Vector<type>::get_size()
+int Vector<type>::get_size()
 {
     return size;
 }
 
 template <class type>
-type Vector<type>::get_value(ll idx)
+type Vector<type>::get_value(int idx)
 {
     assert(idx >= 0 && idx < size);
     return arr[idx];
@@ -53,7 +53,7 @@ type Vector<type>::get_back()
 }
 
 template <class type>
-void Vector<type>::set_value(ll idx, type val)
+void Vector<type>::set_value(int idx, type val)
 {
     assert(idx >= 0 && idx < size);
     arr[idx] = val;
@@ -62,15 +62,15 @@ void Vector<type>::set_value(ll idx, type val)
 template <class type>
 void Vector<type>::print()
 {
-    for (ll i(0); i < size; ++i)
+    for (int i(0); i < size; ++i)
         cout << arr[i] << ' ';
     cout << edl;
 }
 
 template <class type>
-ll Vector<type>::find(type val)
+int Vector<type>::find(type val)
 {
-    for (ll i(0); i < size; ++i)
+    for (int i(0); i < size; ++i)
         if (arr[i] == val)
             return i;
     return -1;
@@ -85,12 +85,12 @@ void Vector<type>::push_back(type val)
 }
 
 template <class type>
-void Vector<type>::insert(ll idx, type val)
+void Vector<type>::insert(int idx, type val)
 {
     assert(0 <= idx && idx < size);
     if (size == capacity)
         expand_capacity();
-    for (ll p(size - 1); p >= idx; --p)
+    for (int p(size - 1); p >= idx; --p)
         arr[p + 1] = arr[p];
     arr[idx] = val;
     ++size;
@@ -100,7 +100,7 @@ template <class type>
 void Vector<type>::right_rotate()
 {
     type last_element(arr[size - 1]);
-    for (ll p(size - 2); p >= 0; --p)
+    for (int p(size - 2); p >= 0; --p)
         arr[p + 1] = arr[p];
     arr[0] = last_element;
 }
@@ -109,13 +109,13 @@ template <class type>
 void Vector<type>::left_rotate()
 {
     type first_element(arr[0]);
-    for (ll p(1); p < size; ++p)
+    for (int p(1); p < size; ++p)
         arr[p - 1] = arr[p];
     arr[size - 1] = first_element;
 }
 
 template <class type>
-void Vector<type>::right_rotate(ll times)
+void Vector<type>::right_rotate(int times)
 {
     times %= size;
     while (times--)
@@ -123,20 +123,20 @@ void Vector<type>::right_rotate(ll times)
 }
 
 template <class type>
-type Vector<type>::pop(ll idx)
+type Vector<type>::pop(int idx)
 {
     assert(idx >= 0 && idx < size);
     type val(arr[idx]);
-    for (ll p(idx + 1); p < size; ++p)
+    for (int p(idx + 1); p < size; ++p)
         arr[p - 1] = arr[p];
     --size;
     return val;
 }
 
 template <class type>
-ll Vector<type>::find_transposition(type val)
+int Vector<type>::find_transposition(type val)
 {
-    for (ll i(0); i < size; ++i)
+    for (int i(0); i < size; ++i)
         if (arr[i] == val)
         {
             if (i == 0)
